Add Generator Interface
Problem
-------

All versions of UUIDs share a common set of bit flags. A general interface is
necessary to to consistently produce flagged UUIDs.

Solution
--------

A generator uses a private writer to fill the byte array with the appropriate
bytes for a given version. The generator then applies all bit flags and returns
the new UUID.