Add Support for Version 1 UUIDs
The RFC-4122 specification calls for time based UUIDs which use a combination of
timestamps, clock sequence, and a node identifier.

This implementation meets the spec by using a 100 nanosecond interval timing
function which deals with auto incrementing values if multiple calls resolve the
so the same value. This timestamp is correctly arranged into the resulting UUID
byte array.

A clock sequence is randomly generated per instance of a generator. This
value is cached and written to the byte array provided by the generator.

A list of interfaces is passed into the constructor of the generator. The first
Interface that has a hardware value is used and written to the byte slice of the
given uuid provided by the generator.